Check if span is still open in ScopeManager.active
The ScopeManager.active and ScopeManager.root function returned spans as
active or as a valid root span when they were closed. This is a
different state from a NoopSpan where no span has been active.

When a span closes it doesn't necessary reset the scope (remove the
executionAsyncId) from the scopes and roots maps on the ScopeManager.

Keep track if a span is open or not by keeping an `open` attribute on
the Span that is set to `false` whenever `Span.close()` is called. We
should not rely on the `isObject() => { closed: true }`, because that's
a more intensive operation (creating and parsing a JSON object) and we
should only use it for testing.

With the new `Span.isOpen()` function we can check if the Span is stil
open and only return it if that's the case. If it's not, also
immediately remove it from the ScopeManager's maps.

I've updated the `createRootSpan` test to close the span after the
assertions, otherwise it will now fail because of returning undefined.

We ran into this issue occurring in PR #669. It will not necessarily fix
the issue we ran into there, but it's an issue nonetheless.