Don't transfer closed spans to new async contexts
This adds another level of protecting to using closed spans in any
context when returned by the ScopeManager.

It's really tricky to test this in isolation, because fetching the
active span, already performs the check. I can't check the private maps
of spans in the ScopeManager itself in the test. I've tested it by
temporarily removing the `&& span.open` check from
`ScopeManager.active()` which confirmed it works.