Check if span is open in ScopeManager.withContext
Before passing through the given span as the current context, make
sure it's still open. If not, do not set it and reuse the existing
active span.

If no span is active, return NoopSpan. Returning a NoopSpan seems like
better behavior than returning a closed span where people can't tell if
it does anything or not.

I've also added a private `setActive` helper that does the same thing as
rootSpan but for active spans. That way it's more centralized how we set
the active and root spans.