Ignore HTTP requests to socket.io
For some reason I haven't quite managed to understand in full, our
HTTP instrumentation breaks socket.io by wrapping one of the event
callbacks in the request object. It is unclear why wrapping it
breaks it -- you can read more about it on [this comment for the
related support issue][descent-into-madness].

This commit adds the default path used by the server-side socket.io
library to the list of ignored paths in the HTTP instrumentation.
While this does not address the underlying issue, it will hopefully
side-step it for most socket.io users.

[descent-into-madness]: https://github.com/appsignal/support/issues/173#issuecomment-1142512153