Copy original function name to wrapper
When wrapping a function in `.bindContext`, the `length` property
of the original function is copied over to the wrapper function.

This commit changes it so that, in addition to the `length`
property, the `name` property is also copied. It also adds a test
that checks the the length and name are copied.