Remove OpenSSL workaround for Next.js
When Node.js version 17 was released, it included an upgrade to
OpenSSL 3.0.0. This version of OpenSSL did not support some cyphers
that were commonly used in Webpack workflows to hash file contents.

A workaround was added then to force Node.js to use an older
OpenSSL version. This means we've been running all our Node.js
tests with an uncommon configuration setting since then.

This commit removes the workaround. Instead, it adds a mechanism
to the build matrix that allows certain variations of Node.js
packages under test to be excluded from certain Node.js versions,
and it uses it so that the Next.js versions whose bundled Webpack
configurations use this cypher are not tested on Node 18.