Don't set closed spans as active in withContext
When a span is given to `ScopeManager.withContext` check if it is is
open. If it's not don't set it as active. Instead pass in the already
currently active span, or a NoopSpan if no span is active.

This prevents the usage of closed spans, that cannot be used to store
any data after they are closed.