Don't restore the root span for withContext
The ScopeManager.withContext only temporary sets the current active
span. It then restores the active span afterwards.

But it also restores the root span, even though that should not have
changed. At least not by anything we've done in this withContext
function.

If the user has set another rootSpan, that is their decision, but it's
not up to us to restore it.