Don't restore active span if closed in withContext
When a span gets closed for whatever reason before getting restored in a
withContext callback, do not restore it. We do not restore closed spans
as active spans.

This is difficult to test, because the `active()` function also does
this check, but removing that check this test will still pass.