Add OpenTelemetry wrapper client
Add a client that disables the auto instrumentation when using this
package for OpenTelemetry. This inherits from the BaseClient but
provides the default instrumentation necessary for better OpenTelemetry
compatibility.

For the user this means they no longer need to configure the
`instrumentHttp`, `instrumentRedis` and `instrumentPg` options manually
anymore. We can update the docs to no longer mention them either for the
OpenTelemetry config. We do need to update the docs to use the new
OpenTelemetry client instead.

```js
// appsignal.js
const { AppsignalOpenTelemetry } = require("@appsignal/nodejs");

const appsignal = new AppsignalOpenTelemetry({
  // Your config
});

module.exports = { appsignal };
```