Runtime dep @opentelemetry/sdk-trace-base package
In PR #659 we tried to import the OpenTelemetry types at compile time
and include them in our own package, but that's not how TypeScript's
`import type` turns out to work.

This breaks the app for non-OpenTelemetry users as reported in #684. The
easiest fix we can ship right now is to add the package as a runtime
dependency rather than only a development dependency.

Ideally though, we do not ship additional dependencies for thing not all
our users use, but we can figure that out after this immediate fix.

I had missed this in PR #659, but we've also already included the
`@opentelemetry/api` as a dependency since PR #651. Already doing the
same this change does for the `@opentelemetry/sdk-trace-base` package.