Improve error reporting on extension load failure
It's a common problem for apps first installed on the host system
(macOS) and then mounted with the `node_modules` directory in a Docker
container that's running another Operating System like Ubuntu to not
start the extension. This is because the extension is not installed for
the container's architecture.

If the target and architecture mismatch, print the installed and current
target and architecture. This will hopefully make the error more clear,
that the user needs to reinstall the extension for their host's
architecture.

I've updated the log message format to match the agreed upon format for
STDERR messages: `appsignal <TYPE>: <message>`
https://github.com/appsignal/integration-guide/blob/main/build/logging.md#warnings-printed-to-stderr
(private link)

I've also added logging for when the install report is not present. This
then returns `unknown-unknown` as the target and architecture. This is
not a pretty log message, but it is such an edge case when the report is
not present at all that I don't think will happen often. It's not worth
the effort to make more pretty right now.

Closes #658