Turn monorepo into regular repo
Change repo structure not to be a monorepo anymore. Now that we are only
relying on the "nodejs" package, there's no need to use workspaces as it
adds friction to the development process, and the idea is to not add
more packages to the same repo in the future.

@tombruijn: I've also updated the build matrix generator to work with
the single package.

Written by: (I'm just reapplying it)

Co-authored-by: Noemi <noemi@appsignal.com>
Co-authored-by: Luismi Ramirez <luismi.ramirez@protonmail.com>