Accept all case diagnose send report confirm input (#711)
We prompt the user with the following when running the diagnose tool:

```
Send diagnostics report to AppSignal? (Y/n):
```

When a user submits this with "Y", we do not accept it. This was because
the input checker only accepted for the lowercase value.

I've lowercased all input, so that "Y", "y", "N" and "n" all work like
they should in this confirmation prompt.

Part of https://github.com/appsignal/support/issues/204