Do not instrument Koa generator middleware
In Koa versions 0.x and 1.x, it was possible to implement
middleware using generator functions. In Koa version 2.x, these are
referred as "legacy middleware" and their use is deprecated,
emitting a warning when they are used. However, it is still
possible to use them, as Koa uses an utility called `koa-convert`
internally to transform legacy middleware into new-style
middleware. In Koa version 3.x, support for generator middleware
will be removed entirely.

Our instrumentation does not support generator-based middleware,
and it will break the request if a generator-based middleware is
used early in the middleware chain, by failing to yield to the
following middleware. Unfortunately, low-quality tutorial websites
continue to recommend using generator-based middleware for error
handling, and these sites rank high for relevant keywords.

This commit changes our Koa instrumentation so that it does not
attempt to instrument generator-based middleware. On Koa version
2.x, this allows the Koa internals to convert the middleware and
emit the deprecation warning. On Koa version 3.x, this will
outright fail, which is the best possible deprecation warning.

The docs are also updated to point out that generator-based
middleware will not be instrumented, and it points users to
`koa-convert`, which they can use to convert the middleware
_before_ passing it to the Koa application's `.use` method,
which would allow AppSignal to instrument it.