Clean the build dir in `clean:ext` script
I was testing the agent's integration tests in the agent repository.
Locally I kept running into the issue that I was testing with the wrong
build of the agent. Removing the `build/` directory seems to help test
against the latest locally compiled version of the agent.

This may need another fix on the agent's side to call this task.

[skip changeset]