Only initialise OpenTelemetry when active
When ininitialising AppSignal, only initialise OpenTelemetry when
active.

Use a ProxyTracerProvider as the exposed `tracerProvider` property
on the client. When initialising OpenTelemetry, set it to delegate
to the initialised NodeTracerProvider. If OpenTelemetry is not
active, the ProxyTracerProvider acts as a NoopTracerProvider.

It is not possible to use NoopTracerProvider directly as it is not
an exported type.