Implement `instrumentations` config option
When initialising AppSignal, it is now possible to pass an
`instrumentations` config option.

This option takes a map as an argument. The keys of the map are
instrumentation package names.

To disable a default instrumentation, pass `{enabled: false}` as
the value.

To change the configuration of a default instrumentation, pass
the desired configuration as the value. The given configuration
will be merged with the default configuration for that
instrumentation.

To add a new instrumentation, pass the initialised instrumentation
as the value.

```
const appsignal = new Appsignal({
  instrumentations: {
    // disable koa instrumentation
    "@opentelemetry/instrumentation-koa": {
      enabled: false
    },
    // change the config for the http instrumentation
    "@opentelemetry/instrumentation-http": {
      ignoreIncomingPaths: [/test/i]
    },
    // add a custom instrumentation
    "@someone/custom-instrumentation": new CustomInstrumentation()
  }
})
```

To provide an upgrade path for adding more default instrumentations
in the future without breaking the application for users who already
added the instrumentation on their own, if an instrumentation
instance is given instead of a config hash, instead of crashing, the
instance is used.