Bump dependencies and add overrides
Bump OpenTelemetry library dependencies as to be able to use
`trace.getActiveSpan()`, which was added in version 1.2.0 of the
`@opentelemetry/api` library. Bump other OpenTelemetry libraries,
as they are all in a weird intricate lockstep with each other.

To avoid type errors where TypeScript complains that two types
from different versions of the same package are different types,
add overrides to force package deduplication.