Add `{set,send}Error` helpers
Adds the `setError` and `sendError` helpers, as described in #741.

Note that the `sendError` callback does not receive a span as an
argument. It creates an active span and runs your callback inside
it. As the other helpers do not require a span as an argument, you
can use those helpers from the callback without having a span
passed as an argument.