Use helpers to set magic attributes
Use our helpers internally when setting errors, params or magic
attributes for the Express and Koa instrumentations. As a result,
magic attributes for Express params and cookies are now set in the
root span, not in the request handler span.

Fix existing tests for the params and add tests for the session and
the error in Express, as well as a test for custom instrumentation.