Warn if there is no active span
Instead of silently failing, emit a debug-level warning to the logs
when a helper attempts to set an attribute, but there is no active
span to set it on.

Also emit a debug-level warning to the logs when a helper that takes
an Error-like object as argument receives an value which is not
Error-like, as it is common to throw non-Error values.