Add root name magic attribute helper
Allow users change the incident's action name by overriding the root
span name for OpenTelemetry traces. This is the counterpart to the
`setName` helper that only changes the active span's name. The "name"
translates to the "title" you see on the box when you hover over an
event in the event timeline.

This helper does the same as `rootSpan.setName("Name")`, but not always
can users access the root span of the trace. Which is where this helper
comes in to make it still possible to overwrite the root span's name. If
the `setRootName` function is used, it overwrites any `rootSpan.setName`
value.

I'm also okay with changing this name to `setTraceName` if that would
make it more clear what is the root in this context.

This also requires the processor change I submitted earlier.

[skip changeset]

Part of https://github.com/appsignal/opentelemetry/issues/46