Add OpenTelemetry SpanKind as a magic attribute
To identify the origin of the HTTP requests from their spans, the
`SpanKind` attribute is needed. As adding a completely new attribute
will add a lot of overhead requiring modifications in all steps, the
attribute is now added from the SpanProcessor as a magic attribute.

More info on OpenTelemetry's SpanKind:
https://opentelemetry.io/docs/reference/specification/trace/api/#spankind