Allow span to be passed to helpers
Not every instrumentation makes the spans that they're working
with into active spans for the current context. Additionally,
some instrumentations provide hooks where the relevant span is
provided as an argument, but is not the active span.

To account for these scenarios, the helpers now take an optional
span argument as their final argument (except for sendError, which
always creates a new span)