Pass span to params helper in Koa hook
When the Koa instrumentation request hook is executed, the current
active span is the span from the _previous_ layer. This means that,
for the router layer, the active span is the last middleware in the
middleware chain.

While this is fine functionally, it makes the behaviour of the hook
hard to reason about. This change makes it so that the params are
set in the span passed to the hook, instead of in the active span.

The integration tests are also amended to check for the params in
the router child span.