Add helpers optional span argument test
In #757, the signature of the helpers was changed to add an optional
span argument. If present, the custom attribute or event set by the
helper will be set in the given span, instead of on the current
active span.

I meant to add tests for it when making the change, but then I
forgot. Here are the tests.