Implement Logger class
Implements a logger class that uses the C API bridge to log messages
through the Logging API. It follows the same user-facing API as the
one implemented in the [Ruby logging PR][pr], except that it uses
the logging levels that were in use in the Node.js logger.

On the user facing API, the logger level is implemented as a string.
The attributes are typed as a flat map whose values must be strings,
numbers or booleans, in concordance with how the agent traverses the
resulting data map.

[pr]: https://github.com/appsignal/appsignal-ruby/pull/890