Move integration test config helpers to module
Instead of repeating the `ENV.fetch(...)` code everywhere the
environment variables are read, move this logic to a helper.

Whenever we change the ENV var name or add additional checks to this
value, it will be easier to update.

I also added some defaults, so it's easier to run locally. Those
defaults are for running it locally and do not match the Docker test
env.

I wanted to quickly run one of the test apps to see what kind of span
attributes are available for a GraphQL span.

Related to https://github.com/appsignal/appsignal-nodejs/issues/792