Set RuboCop Ruby target version
Update Rubocop to work better with newer Ruby versions.

Configure a default Ruby version for the project and RuboCop so we're
all working with the same Ruby versions. RuboCop and Semaphore will both
read from the `.ruby-version` file.

Without this I got errors running it on a newer Ruby version.

```
An error occurred while Layout/BlockAlignment cop was inspecting
  appsignal-nodejs/test/express-mongoose/tests/spec/app_spec.rb:3:0.
```

[skip changeset]