Add version 3 release section to changelog
Rather than just a collection of different pre-releases that together
form the version 3 changelog, add a new section for the release. In this
section briefly summarize the new additions and most important changes.
The other changes can still be found in the pre-release changelogs.

I would recommend merging this just before release, because if we
merge it and then do more prereleases, the ordering will be odd for the
reader.

[skip ci]

Closes #794