Add noop user and integration logger
This makes it safe for our code to call `Client.integrationLogger`
if the client is not active or is not initialised. For example, when
running scripts that do not `--require './appsignal.cjs'` (such as
test scripts) but that may call code that indirectly attempts to use
the integration logger (such as the custom instrumentation helpers)

It also makes it safe for users to call `Client.logger("foo")` and
call methods on that logger when the client is not active or is
not initialized. This allows users who integrate our AppSignal
logger into their code to use that same code in contexts where
AppSignal is not active or initialised.