Track restify request parameters
Upgrade to the latest restify package that includes the `requestHook`
config option to report the parameters.

We only track them for the request handler layer type and not all
middlewares that may be present to avoid it being set multiple times and
it overwriting one another.

This relies on the `params` and `query` properties on the request. These
are the parsed request params and query params for the request that are
only available when the restify queryParser plugin is loaded. This
avoids us having to parse it separately, and I assume it will only be
necessary to track parameters if the plugin is added to the app.