Add warnings to 2.x old packages
Avoid confusion when people view our 2.x packages on npm when they're
using 3.x or newer. Add a small warning to the top of the README to link
people to the v3 docs.

This will require a patch release for all packages to have this update
show up on npmjs.org.

Closes #869