Adapt HTTP and Postgres span expectations
The HTTP spans no longer start with "HTTP" (so, instead of "HTTP GET",
it just says "GET") and the HTTP spans wrapping the Apollo and Yoga
spans now have `POST /graphql` in their span name. This seems like a
good change.

The PostgreSQL spans now have the name of the DB after the span
name, and their name more closely resembles a span category. In
addition, there are now spans that represent the connection. We
should amend the extractor to better deal with these spans, but it's
okay for now.