Fix musl platform comparison
When the extension is installed in musl Linux, but the extension
fails to load for some other reason, the error displayed would
describe a bogus platform mismatch between musl and non-musl Linux.

Fix this by checking whether the current platform is musl Linux,
and expecting the corresponding platform target name to be found in
the install report. Do so by re-using the logic from the installation
report helpers.