Reduce GraphQL instrumentation noise
This configuration should greatly reduce the amount of noise the
GraphQL instrumentation produces on the event timeline.

`mergeItems` merges resolvers for fetching multiple elements from
a list into one, `ignoreTrivialResolveSpans` skips resolvers that
just transparently map to an attribute (which contribute a ton of
noise, and `allowValues` (set to false, which is the default)
ensures that the GraphQL query is sanitized.