Fix Next.js test app assertion
Updating Next.js adds the path of the route to the span name. This isn't
a problem for us. We only use the span name as a fallback, and our agent
tests already assume it can contain the route path.

It also now is reported by the `next.js` instrumentation, rather than
the default http one, but we need it to report one from the `next.js`
integration, so I'm not sure how it worked before.