Implement `opentelemetryInstrumentations` getter
Implement an `opentelemetryInstrumentations` getter on the client,
to allow users of the `initializeOpentelemetrySdk` config option to
configure OpenTelemetry in an AppSignal-compatible way.

This getter honors the settings in the `disableDefaultInstrumentations`
and `additionalInstrumentations` configuration options. Implementing
this was a bit of a refactor, so I took a moment to add tests for this
and the `initializeOpentelemetrySdk` config option.