Fix Express error handler signature error
Fixed an error thrown when using the Express error handler that depended
on which types package version for Express is used.

We've tried to fix the types but couldn't get to make it work due
to how the interfaces in `@types/express-serve-static-core` are. We
tried to replicate the exact same types allowed in the `IRouterHandler`
interface without any success.

The return type of the error handler cannot be an `ExpressErrorHandler`
anymore as it doesn't comply with the latest types defined by the
packages.

This commit removes the return type of the error handler and makes the
`req` and `res` argument to be of `any` type to stop the error from
happening and to hopefully keep the handler compatible in future
possible patches of the types packages.

Co-authored-by: Noemi Lapresta <noemi@appsignal.com>
Co-authored-by: Luismi Ramirez <luismi@appsignal.com>