Add setSqlBody tracing helper
Make it easier to set SQL queries as the span body attribute without
danger of storing unsanitized SQL queries.

When `setSqlBody` is used in combination with `setBody`, the
`setSqlBody` call is leading.

Related to https://github.com/appsignal/appsignal-python/issues/140

This change was made available in the agent in PR #955

Part of #956