Force usage of `@opentelemetry/api` version 1.6
Accidentally using several `@opentelemetry/api` minor versions,
which is almost inevitable due to how npm resolves dependencies,
triggers an absolutely deranged "feature" in the OpenTelemetry API
internals, by which the global tracer set by the OpenTelemetry API
version with the lower version number will not be visible by the
one with the higher version number, and an error message about it
will be logged to nowhere.

This is not helped by the fact that `@prisma/instrumentation` has a
hardcoded version bound, which they bump automatically for no reason
whatsoever.

Do a best-effort attempt to force OpenTelemetry versions to a
specific version range instead, by picking versions of the
instrumentations that are compatible with the `@opentelemetry/api`
minor version. This will hopefully trick OpenTelemetry into picking
versions that it can dedupe. Maybe. Who knows.

We'll have to remember to update this in the future, whenever the
release tracker tells us to.

If this breaks again in the future, let's tighten the version bounds
further.