Rerun npm install to rebuild package-lock.json
For some reason in the merge commit with the main branch it installed
the wrong versions of packages and I got a linter issue on one of the
files in the package. Rebuild the `package-lock.json` file (by removing
`node_modules` and `package-lock.json`, and running `npm install`) and
commit the result.