Fix demo sample recognition by fixing the demo tag
When `demo_sample` was passed as an attribute to the
`createOpenTelemetrySpan` extension function it wasn't recognized as a
tag for AppSignal. The UI only recognizes tags and not root level
attributes for the demo samples box.

Use the attribute format described in our docs to transform the
attribute into a tag.
https://docs.appsignal.com/opentelemetry/custom-instrumentation/attributes.html#tag