Fix boolean config options set to false
When a boolean config option was set to false when initializing the
client and it wasn't its default value, it wouldn't work. This was due
to they way we checked the existence of values when setting the private
env vars. As it was a regular `if`, a `false` value would always be
`false` and the transformation skipped.

This commit checks if the present config value is of boolean type and
transforms it into a string nevermind its value.