Await heartbeat promises on shutdown
When AppSignal is "gracefully stopped" by calling `Appsignal.stop()`,
the heartbeats might not have finished sending. AppSignal should await
the delivery of those heartbeats before shutting down.

To do this, the pending heartbeats are collected into a set, which
automatically cleans itself up when all promises are settled. A promise
for the settling of all remaining promises in the set is awaited when
`Appsignal.stop()` is called.

The method that awaits these promises is named `Heartbeat.shutdown()`,
and not `Heartbeat.stop()`, to avoid any potential confusion where it
could be mistaken for the counterpart of `Heartbeat.start()`.

This is a sort of breaking change, in that `Appsignal.stop()` is now
an asynchronous method, meaning that it returns a promise, and as
such it must be awaited.

However, note that `Appsignal.stop()` should have already been
awaiting the promise returned by the OpenTelemetry SDK's `.shutdown()`
method. As such, this change makes sense beyond the heartbeats.