Support async functions
In practice, most long-running operations in Node.js applications
take place as a sequence of `await` operations on functions that
return promises, within an `async` function that itself returns
a promise.

As such, in order for AppSignal to be able to track the duration
of the heartbeat, when the return value is a promise, AppSignal
should send the heartbeat when the promise is resolved, and not when
the function that returns the promise finishes.