Allow manually awaiting heartbeat events
In addition to the easy-to-use `heartbeat` function, the `Heartbeat`
class is also exposed, allowing finer control to customers who might
want to send the `start` and `finish` events independently from each
other.

Since the events are sent asynchronously, it makes sense to return
the corresponding promise so that it can be awaited by the caller.

However, since awaiting a rejected promise raises an error, and we
strongly avoid causing AppSignal-related errors to be raised in
customers' applications, the promise is overriden on rejection with
an empty resolved promise, making it safe to await it.