Write tests for heartbeats
Test the expected behaviour of the heartbeat module against mocked
Nock requests.

Return the same promise from `Appsignal.heartbeat()` that was
returned by the function given to it as an argument, instead of
returning a wrapper promise that emits the heartbeat.

Fix a bug where the timestamp was sent in milliseconds instead of
seconds.