Update OpenTelemetry package dependencies
Update the OpenTelemetry dependencies to fix issue #1068, about the
types package for ioredis not being present anymore. I've updated
everything up to OpenTelemetry API 1.8. Version 1.9 is also out, but the
prisma package is not compatible with that at the moment.

Lock all dependencies to a minor version range so that it hopefully
doesn't install different versions of the OpenTelemetry API package,
which breaks things.

Related commit with more information:
32c499fa0c07535f45c940df1f3b8ab3aa3822b8
Closes #1068