Avoid probes interval blocking shutdown
The probes instance starts an interval timer on initialisation.
By default, timers cause the Node.js engine to await their
completion on shutdown -- which, this being an interval timer, is
never actually completed. (See #418)

We fixed this in the past by implementing `Appsignal.stop`, which
clears the interval. But a preferable solution is to mark the
probes' interval as `.unref()`, telling the Node.js engine that it
does not need to be awaited.