Remove redundant cron check-in pairs
See https://github.com/appsignal/appsignal-ruby/pull/1407 for context.
This commit reimplements that in Node.js.

Before submitting a batch of cron check-in events, check for any pairs
of events that are redundant with other pairs of cron check-in events.
This would happen, for example, when running a `cron` block for a given
identifier many times in quick succession.

A redundant pair of cron check-in events is a pair of start and
finish cron check-in events for a given identifier, given that another
pair of start and finish check-in events exists for that identifier.