Disable distributed OpenTelemetry propagation
Prevent OpenTelemetry from sending tracing headers to other applications
through HTTP requests.

Use an unconfigured `CompositePropagator` to disable HTTP
propagation headers while maintaining local trace correlation.
It is not configured with any other propagators, so it will not add the
headers to the outgoing requests.

I hope this fixes the issue with app data being reported to the wrong
app, if one app makes a request to another app and they're both
instrumented with OpenTelemetry.